% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenWeatherMap.R
\name{get_current_weather}
\alias{get_current_weather}
\title{get current weather data for one location}
\usage{
get_current_weather(api_key, cityID = NA, city = "", country = "",
  coordinates = NA, zip_code = NA)
}
\arguments{
\item{api_key}{Open weather map API key}

\item{cityID}{city ID}

\item{city}{name of city}

\item{country}{name of country}

\item{coordinates}{(lat,lon) coordinates of the location of your interest}

\item{zip_code}{zip code}
}
\value{
data frame giving current weather data for one location
}
\description{
get current weather data for one location
}
\examples{
\dontrun{
data=get_current_weather(api_key,city="guwahati")
}
}

