% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenWeatherMap.R
\name{get_weather_forecast}
\alias{get_weather_forecast}
\title{get weather forecast data for one location}
\usage{
get_weather_forecast(api_key, cityID = NA, city = "", country = "",
  coordinates = NA)
}
\arguments{
\item{api_key}{Open weather map API key}

\item{cityID}{city ID}

\item{city}{name of city}

\item{country}{name of country}

\item{coordinates}{(lat,lon) coordinates of the location of your interest}
}
\value{
data frame giving weather forecast data for one location
}
\description{
get weather forecast data for one location
}
\examples{
data=get_weather_forecast(api_key,city="guwahati")
}

