% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getCurrentWeatherByCityName}
\alias{getCurrentWeatherByCityName}
\title{Get current weather information}
\usage{
getCurrentWeatherByCityName(place)
}
\arguments{
\item{place}{name of place}
}
\value{
weather information(list)
}
\description{
Get current weather information
}

